+++ packages/forms/src/UIForm/UIForm.component.js
@@ -40,15 +40,16 @@ export class UIFormComponent extends React.Component {
 		this.setFormRef = this.setFormRef.bind(this);
 		// control the tv4 language here.
 		const language = getLanguage(props.t);
+		const languageId = props.id || '@talend';
 		if (props.language != null) {
 			Object.assign(language, props.language);
-			// Force update of language @talend even if already set
-			tv4.addLanguage('@talend', language);
-			tv4.language('@talend');
+			// Force update of language by id even if already set
+			tv4.addLanguage(languageId, language);
+			tv4.language(languageId);
 		}
-		if (!tv4.language('@talend')) {
-			tv4.addLanguage('@talend', language);
-			tv4.language('@talend'); // set it
+		if (!tv4.language(languageId)) {
+			tv4.addLanguage(languageId, language);
+			tv4.language(languageId); // set it
 		}
 		const allFormats = Object.assign(customFormats(props.t), props.customFormats);
 		tv4.addFormat(allFormats);